Initial version of OpenFOAM Horizon Dashboard
Integrates all of the componsent of the OpenFOAM Cloud built on top of
OpenStack Horizon. There are two panels available. Simulations represent
a high-level description of all simulations that are actively running by
the current user/tenant. Each simulation consists of one or more
instances which can be seen in the instances panel.

Code cleansing is required!

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>